[core] separate network-indicator and only fetch blocknumber for checking
