+++ src/renderer/coreapp/accounts-bar.tsx
@@ -1,31 +1,17 @@
-import * as React from 'react'
-import BigNumber from 'bignumber.js'
+import { Account } from '../../lib/accounts'
+import { fmtAddress } from '../../lib/utils'
 
+import * as React from 'react'
 import { makeStyles } from '@material-ui/core/styles'
 import Select from '@material-ui/core/Select'
 import MenuItem from '@material-ui/core/MenuItem'
 import Box from '@material-ui/core/Box'
 import Typography from '@material-ui/core/Typography'
+
 import { AddressOnlyAccountIcon, LedgerAccountIcon, LocalAccountIcon } from './accounts-app/account-icons'
-import WifiIcon from '@material-ui/icons/Wifi'
-import WifiOffIcon from '@material-ui/icons/WifiOff'
+import NetworkIndicator from './network-indicator'
 
-import { Account } from '../../lib/accounts'
-import { fmtAddress } from '../../lib/utils'
-import kit, { useNetworkURL } from '../state/kit'
-import { CFG, networkName } from '../../lib/cfg'
-import Tooltip from '@material-ui/core/Tooltip'
-import Button from '@material-ui/core/Button'
-import Dialog from '@material-ui/core/Dialog'
-import DialogTitle from '@material-ui/core/DialogTitle'
-import DialogContent from '@material-ui/core/DialogContent'
-import DialogActions from '@material-ui/core/DialogActions'
-import TextField from '@material-ui/core/TextField'
-import { newKit } from '@celo/contractkit'
-import LinearProgress from '@material-ui/core/LinearProgress'
-import { UserError } from '../../lib/error'
-
-const useStyles = makeStyles((theme) => ({
+const useStyles = makeStyles(() => ({
 	name: {
 		width: 120,
 		fontFamily: "monospace",
@@ -36,12 +22,6 @@ const useStyles = makeStyles((theme) => ({
 	address: {
 		fontFamily: "monospace",
 	},
-	connected: {
-		color: theme.palette.success.main,
-	},
-	disconnected: {
-		color: theme.palette.error.main,
-	},
 }))
 
 const AccountsBar = (props: {
@@ -50,57 +30,10 @@ const AccountsBar = (props: {
 	onSelectAccount: (a: Account) => void,
 }): JSX.Element => {
 	const classes = useStyles()
-	const [connected, setConnected] = React.useState(true)
-	const maxBlockDelaySecs = 30000
-	React.useEffect(() => {
-		let errCnt = 0
-		const timer = setInterval(async () => {
-			const k = kit()
-			try {
-				const block = await k.web3.eth.getBlock('latest')
-				if (!block || block.number <= 1) {
-					throw new Error(`No latest block?`)
-				}
-				errCnt = 0
-				const delaySecs = Date.now() / 1000 - new BigNumber(block.timestamp).toNumber()
-				setConnected(delaySecs <= maxBlockDelaySecs)
-			} catch (e) {
-				errCnt += 1
-				if (errCnt >= 2) {
-					setConnected(false)
-				}
-			}
-		}, maxBlockDelaySecs/6)
-		return () => { clearInterval(timer) }
-	}, [])
-	const [networkURL, setNetworkURL] = useNetworkURL()
-	const [openNetworkURL, setOpenNetworkURL] = React.useState(false)
-
-	const handleOpenNetworkURL = () => { setOpenNetworkURL(true) }
-	const handleCloseNetworkURL = (v: string) => {
-		if (v !== networkURL) {
-			setNetworkURL(v)
-			setConnected(true)
-		}
-		setOpenNetworkURL(false)
-	}
-	const netName = networkName(CFG().networkId)
 	return (
 		<Box display="flex" flexDirection="row" justifyContent="flex-end" p={2}>
-			{openNetworkURL &&
-			<ChangeNetworkURL
-				networkURL={networkURL}
-				onClose={handleCloseNetworkURL}
-				/>}
 			<Box display="flex" flexDirection="row" flex={1}>
-				<Tooltip title={networkURL}>
-					<Button
-						endIcon={connected ?
-							<WifiIcon className={classes.connected} /> :
-							<WifiOffIcon className={classes.disconnected} />}
-						onClick={handleOpenNetworkURL}
-						>{netName}</Button>
-				</Tooltip>
+				<NetworkIndicator />
 			</Box>
 			<Select
 				value={props.selectedAccount?.address || ""}
@@ -131,61 +64,4 @@ const AccountsBar = (props: {
 		</Box>
 	)
 }
-
-const ChangeNetworkURL = (props: {
-	networkURL: string
-	onClose: (v: string) => void
-}) => {
-	const [networkURL, setNetworkURL] = React.useState(props.networkURL)
-	const [isTesting, setIsTesting] = React.useState(false)
-	const onClose = props.onClose
-	React.useEffect(() => {
-		if (!isTesting) {
-			return
-		}
-		const kit = newKit(networkURL)
-		kit.web3.eth.net
-			.getId()
-			.then((networkId) => {
-				const cfgNetworkId = CFG().networkId
-				if (networkId.toString() !== cfgNetworkId) {
-					throw new UserError(`NetworkId doesn't match. Expected: ${cfgNetworkId}, Got: ${networkId}.`)
-				} else {
-					onClose(networkURL)
-				}
-			})
-			.catch((e) => {
-				setIsTesting(false)
-				throw e
-			})
-			.finally(() => {
-				kit.stop()
-			})
-	}, [onClose, isTesting, networkURL])
-	const handleCancel = () => { onClose(props.networkURL) }
-	const handleConnect = () => { setIsTesting(true) }
-	return (
-		<Dialog open={true} onClose={handleCancel}>
-			<DialogTitle>Change Network</DialogTitle>
-			<DialogContent style={{minWidth: 500}}>
-				<TextField
-					autoFocus
-					margin="dense"
-					label={`Network URL`}
-					value={networkURL}
-					size="medium"
-					fullWidth={true}
-					onChange={(e) => { setNetworkURL(e.target.value) }}
-				/>
-				<LinearProgress
-					style={{visibility: !isTesting ? "hidden" : undefined}} color="primary" />
-			</DialogContent>
-			<DialogActions>
-				<Button onClick={handleCancel}>Cancel</Button>
-				<Button onClick={handleConnect} disabled={isTesting}>Connect</Button>
-			</DialogActions>
-		</Dialog>
-	)
-}
-
 export default AccountsBar
